moved extension logic in back to  namespace to avoid breaking iot
