+++ src/azure-cli-core/azure/cli/core/extension/_homebrew_patch.py
