+++ src/azure-cli-core/azure/cli/core/extension/tests/latest/data/my_second_cli_extension-0.0.1+dev-py2.py3-none-any.whl
