+++ src/azure-cli-core/azure/cli/core/extension/tests/latest/data/myfirstcliextension-0.0.3+dev-py2.py3-none-any.whl
