+++ src/azure-cli-core/azure/cli/core/extension/tests/latest/test_index_get.py
@@ -7,7 +7,7 @@ import mock
 import unittest
 from requests.exceptions import ConnectionError, HTTPError
 from azure.cli.core.util import CLIError
-from azure.cli.core.extensions._index import (get_index, get_index_extensions, DEFAULT_INDEX_URL,
+from azure.cli.core.extension._index import (get_index, get_index_extensions, DEFAULT_INDEX_URL,
                                               ERR_TMPL_NON_200, ERR_TMPL_NO_NETWORK, ERR_TMPL_BAD_JSON,
                                               ERR_UNABLE_TO_GET_EXTENSIONS)
 
@@ -75,12 +75,12 @@ class TestExtensionIndexGet(unittest.TestCase):
         with mock.patch('requests.get', side_effect=mock_index_get_generator(DEFAULT_INDEX_URL, data)):
                 self.assertEqual(get_index_extensions().get('myext'), obj)
 
-        with mock.patch('azure.cli.core.extensions._index.logger.warning', autospec=True) as logger_mock:
+        with mock.patch('azure.cli.core.extension._index.logger.warning', autospec=True) as logger_mock:
             with mock.patch('requests.get', side_effect=mock_index_get_generator(DEFAULT_INDEX_URL, {})):
                 self.assertEqual(get_index_extensions(), None)
                 logger_mock.assert_called_once_with(ERR_UNABLE_TO_GET_EXTENSIONS)
 
-        with mock.patch('azure.cli.core.extensions._index.logger.warning', autospec=True) as logger_mock:
+        with mock.patch('azure.cli.core.extension._index.logger.warning', autospec=True) as logger_mock:
             with mock.patch('requests.get', side_effect=mock_index_get_generator(DEFAULT_INDEX_URL, {'v2extensions': []})):
                 self.assertEqual(get_index_extensions(), None)
                 logger_mock.assert_called_once_with(ERR_UNABLE_TO_GET_EXTENSIONS)
