+++ src/azure-cli-core/setup.py
@@ -109,7 +109,7 @@ setup(
         'azure.cli',
         'azure.cli.core',
         'azure.cli.core.commands',
-        'azure.cli.core.extensions',
+        'azure.cli.core.extension',
         'azure.cli.core.profiles',
     ],
     install_requires=DEPENDENCIES,
