+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/custom.py
@@ -1735,7 +1735,7 @@ def _update_addons(cmd, instance, subscription_id, resource_group_name, addons, 
 def _get_azext_module(extension_name, module_name):
     try:
         # Adding the installed extension in the path
-        from azure.cli.core.extensions.operations import add_extension_to_path
+        from azure.cli.core.extension.operations import add_extension_to_path
         add_extension_to_path(extension_name)
         # Import the extension module
         from importlib import import_module
@@ -1780,7 +1780,7 @@ def _handle_addons_args(cmd, addons_str, subscription_id, resource_group_name, a
 
 def _install_dev_spaces_extension(extension_name):
     try:
-        from azure.cli.core.extensions import operations
+        from azure.cli.core.extension import operations
         operations.add_extension(extension_name=extension_name)
     except Exception:  # nopa pylint: disable=broad-except
         return False
@@ -1788,9 +1788,9 @@ def _install_dev_spaces_extension(extension_name):
 
 
 def _update_dev_spaces_extension(extension_name, extension_module):
-    from azure.cli.core.extensions import ExtensionNotInstalledException
+    from azure.cli.core.extension import ExtensionNotInstalledException
     try:
-        from azure.cli.core.extensions import operations
+        from azure.cli.core.extension import operations
         operations.update_extension(extension_name=extension_name)
         operations.reload_extension(extension_name=extension_name)
     except CLIError as err:
@@ -1806,7 +1806,7 @@ def _update_dev_spaces_extension(extension_name, extension_module):
 
 
 def _get_or_add_extension(extension_name, extension_module, update=False):
-    from azure.cli.core.extensions import (ExtensionNotInstalledException, get_extension)
+    from azure.cli.core.extension import (ExtensionNotInstalledException, get_extension)
     try:
         get_extension(extension_name)
         if update:
