+++ src/command_modules/azure-cli-iot/azure/cli/command_modules/iot/__init__.py
@@ -8,7 +8,7 @@ from knack.log import get_logger
 from azure.cli.core import AzCommandsLoader
 from azure.cli.core.commands import CliCommandType
 import azure.cli.command_modules.iot._help  # pylint: disable=unused-import
-from azure.cli.core.extensions import extension_exists
+from azure.cli.core.extension import extension_exists
 
 
 def handler(ctx, **kwargs):
